feat(skeleton): add customizable animation and theme registration
