+++ crates/freya-components/src/theming/themes.rs
@@ -1,3 +1,5 @@
+use std::time::Duration;
+
 use freya_core::prelude::*;
 #[cfg(feature = "titlebar")]
 use torin::prelude::Length;
@@ -48,6 +50,10 @@ use crate::{
     },
     select::SelectThemePreference,
     sidebar::SideBarItemThemePreference,
+    skeleton::{
+        SkeletonAnimation,
+        SkeletonThemePreference,
+    },
     slider::SliderThemePreference,
     switch::{
         SwitchColorsThemePreference,
@@ -613,6 +619,20 @@ fn register_base_component_themes(theme: &mut Theme) {
             corner_radius: Preference::Specific(CornerRadius::new_all(0.0)),
             width: Preference::Specific(Size::Pixels(Length::new(46.0))),
             height: Preference::Specific(Size::Fill),
+        },
+    );
+    let colors = theme.colors.clone();
+    theme.set(
+        "skeleton",
+        SkeletonThemePreference {
+            background: Preference::Reference("surface_primary"),
+            shimmer_color: Preference::Specific(colors.text_placeholder.with_a(90)),
+            duration: Preference::Specific(Duration::from_millis(1000)),
+            animation: Preference::Specific(SkeletonAnimation::Pulse),
+            corner_radius: Preference::Specific(CornerRadius::new_all(4.)),
+            shimmer_from: Preference::Specific(-250.),
+            shimmer_to: Preference::Specific(900.),
+            shimmer_width: Preference::Specific(200.),
         },
     );
 }
