+++ crates/freya/Cargo.toml
@@ -43,7 +43,6 @@ all = [
   "terminal",
   "query",
   "code-editor",
-  "skeletons",
 ]
 all-publish = [
   "all",
@@ -93,7 +92,6 @@ webview = ["dep:freya-webview"]
 titlebar = ["freya-components/titlebar"]
 terminal = ["dep:freya-terminal"]
 code-editor = ["dep:freya-code-editor"]
-skeletons = ["dep:freya-skeletons"]
 
 [dependencies]
 freya-core = { workspace = true }
@@ -114,7 +112,6 @@ freya-query = { workspace = true, optional = true }
 freya-webview = { workspace = true, optional = true }
 freya-terminal = { workspace = true, optional = true }
 freya-code-editor = { workspace = true, optional = true }
-freya-skeletons = { workspace = true, optional = true }
 
 # Performance
 freya-performance-plugin = { workspace = true }
